backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Caching
redis==5.0.1
hiredis==2.3.2

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
chardet==5.2.0

# NLP & ML - Updated for Python 3.12 compatibility
sentence-transformers>=2.3.1
transformers>=4.37.0
spacy>=3.7.2
torch>=2.2.0
scikit-learn>=1.4.0
xgboost>=2.0.3
nltk>=3.8.1

# Keyword Extraction
keybert>=0.8.4
yake>=0.4.8
rake-nltk>=1.0.6

# YouTube API
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
isodate>=0.6.1

# Data Processing
pandas>=2.2.0
numpy>=1.26.3

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
